refactor: on Run scenario without modal
