--- app/layout/scenarios/edit/_run/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
