+++ app/layout/scenarios/edit/parameters/component.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState, useEffect } from 'react';
+import React, { useCallback, useEffect } from 'react';
 
 import { useSelector, useDispatch } from 'react-redux';
 
@@ -7,20 +7,22 @@ import { useRouter } from 'next/router';
 import { getScenarioEditSlice } from 'store/slices/scenarios/edit';
 
 import { AnimatePresence, motion } from 'framer-motion';
+import { usePlausible } from 'next-plausible';
 import { ScenarioSidebarTabs, ScenarioSidebarSubTabs } from 'utils/tabs';
 
+import { useMe } from 'hooks/me';
 import { useProjectRole } from 'hooks/project-users';
-import { useScenario } from 'hooks/scenarios';
+import { useProject } from 'hooks/projects';
+import { useScenario, useSaveScenario, useRunScenario } from 'hooks/scenarios';
+import { useToasts } from 'hooks/toast';
 
 import HelpBeacon from 'layout/help/beacon';
 import Pill from 'layout/pill';
-import Run from 'layout/scenarios/edit/_run';
 import AdvancedSettings from 'layout/scenarios/edit/parameters/advanced-settings';
 import BLMCalibration from 'layout/scenarios/edit/parameters/blm-calibration';
 import Sections from 'layout/sections';
 
 import Button from 'components/button';
-import Modal from 'components/modal';
 
 const SECTIONS = [
   {
@@ -39,10 +41,14 @@ export interface ScenariosSidebarEditAnalysisProps {
 }
 
 export const ScenariosSidebarEditAnalysis: React.FC<ScenariosSidebarEditAnalysisProps> = () => {
-  const [runOpen, setRunOpen] = useState(false);
+  const { addToast } = useToasts();
+  const plausible = usePlausible();
   const { query } = useRouter();
   const { pid, sid } = query;
 
+  const { user } = useMe();
+
+  const { data: projectData } = useProject(pid);
   const { data: projectRole } = useProjectRole(pid);
   const VIEWER = projectRole === 'project_viewer';
 
@@ -54,6 +60,16 @@ export const ScenariosSidebarEditAnalysis: React.FC<ScenariosSidebarEditAnalysis
   const dispatch = useDispatch();
 
   const { data: scenarioData } = useScenario(sid);
+  const { metadata } = scenarioData || {};
+  const { scenarioEditingMetadata, marxanInputParameterFile: runSettings } = metadata || {};
+
+  const saveScenarioMutation = useSaveScenario({
+    requestConfig: {
+      method: 'PATCH',
+    },
+  });
+
+  const runScenarioMutation = useRunScenario({});
 
   useEffect(() => {
     if (!SECTIONS.find((s) => s.id === subtab)) {
@@ -65,6 +81,90 @@ export const ScenariosSidebarEditAnalysis: React.FC<ScenariosSidebarEditAnalysis
     const sub = s || null;
     dispatch(setSubTab(sub));
   }, [dispatch, setSubTab]);
+
+  const onRunScenario = useCallback(() => {
+    const data = {
+      numberOfRuns: runSettings.NUMREPS,
+      boundaryLengthModifier: runSettings.BLM,
+      metadata: {
+        marxanInputParameterFile: runSettings,
+        scenarioEditingMetadata: {
+          ...scenarioEditingMetadata,
+          lastJobCheck: new Date().getTime(),
+          tab: ScenarioSidebarTabs.SOLUTIONS,
+          subtab: null,
+          status: {
+            'protected-areas': 'draft',
+            features: 'draft',
+            analysis: 'draft',
+            solutions: 'draft',
+          },
+        },
+      },
+    };
+
+    saveScenarioMutation.mutate({ id: `${sid}`, data }, {
+      onSuccess: () => {
+        runScenarioMutation.mutate({ id: `${sid}` }, {
+          onSuccess: ({ data: { data: s } }) => {
+            addToast('run-start', (
+              <>
+                <h2 className="font-medium">Success!</h2>
+                <p className="text-sm">Run started</p>
+              </>
+            ), {
+              level: 'success',
+            });
+            console.info('Scenario name saved succesfully', s);
+
+            plausible('Run scenario', {
+              props: {
+                userId: `${user.id}`,
+                userEmail: `${user.email}`,
+                projectId: `${pid}`,
+                projectName: `${projectData.name}`,
+                scenarioId: `${sid}`,
+                scenarioName: `${scenarioData?.name}`,
+              },
+            });
+          },
+          onError: () => {
+            addToast('error-run-start', (
+              <>
+                <h2 className="font-medium">Error!</h2>
+                <p className="text-sm">Scenario name not saved</p>
+              </>
+            ), {
+              level: 'error',
+            });
+          },
+        });
+      },
+      onError: () => {
+        addToast('error-scenario-name', (
+          <>
+            <h2 className="font-medium">Error!</h2>
+            <p className="text-sm">Scenario name not saved</p>
+          </>
+        ), {
+          level: 'error',
+        });
+      },
+    });
+  }, [
+    pid,
+    sid,
+    saveScenarioMutation,
+    runScenarioMutation,
+    addToast,
+    scenarioEditingMetadata,
+    plausible,
+    projectData?.name,
+    user?.email,
+    user?.id,
+    scenarioData?.name,
+    runSettings,
+  ]);
 
   if (!scenarioData || tab !== ScenarioSidebarTabs.PARAMETERS) return null;
 
@@ -158,7 +258,7 @@ export const ScenariosSidebarEditAnalysis: React.FC<ScenariosSidebarEditAnalysis
                   theme="spacial"
                   size="lg"
                   disabled={VIEWER}
-                  onClick={() => setRunOpen(true)}
+                  onClick={onRunScenario}
                 >
                   Run scenario
                 </Button>
@@ -166,14 +266,14 @@ export const ScenariosSidebarEditAnalysis: React.FC<ScenariosSidebarEditAnalysis
             )}
           </AnimatePresence>
 
-          <Modal
+          {/* <Modal
             title="Run scenario"
             open={runOpen}
             size="narrow"
             onDismiss={() => setRunOpen(false)}
           >
             <Run />
-          </Modal>
+          </Modal> */}
         </motion.div>
       </HelpBeacon>
     </div>
